+++ CHANGELOG.md
@@ -40,6 +40,7 @@
    command has been renamed `autodetect`.
  - Changed `luakit://introspector/` to `luakit://binds/`.
  - URL completion now uses word-based fuzzy matching.
+ - `:download` now uses the current page URI by default.
 
 ### Removed